Splitted flags series into several files.
