+++ js/Series/Flags/FlagsPoint.js
@@ -1,0 +1,67 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+'use strict';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../Core/Series/Series.js';
+var ColumnSeries = BaseSeries.seriesTypes.column;
+import U from '../../Core/Utilities.js';
+var isNumber = U.isNumber;
+/* *
+ *
+ *  Class
+ *
+ * */
+var FlagsPoint = /** @class */ (function (_super) {
+    __extends(FlagsPoint, _super);
+    function FlagsPoint() {
+        /* *
+         *
+         *  Properties
+         *
+         * */
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        _this.options = void 0;
+        _this.series = void 0;
+        return _this;
+    }
+    /* *
+     *
+     *  Functions
+     *
+     * */
+    /* eslint-disable valid-jsdoc */
+    /**
+     * @private
+     */
+    FlagsPoint.prototype.isValid = function () {
+        // #9233 - Prevent from treating flags as null points (even if
+        // they have no y values defined).
+        return isNumber(this.y) || typeof this.y === 'undefined';
+    };
+    return FlagsPoint;
+}(ColumnSeries.prototype.pointClass));
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default FlagsPoint;
