+++ ts/Series/Flags/FlagsPoint.ts
@@ -1,0 +1,89 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+'use strict';
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ColorType from '../../Core/Color/ColorType';
+import type CSSObject from '../../Core/Renderer/CSSObject';
+import type FlagsPointOptions from './FlagsPointOptions';
+import type FlagsSeries from './FlagsSeries';
+import BaseSeries from '../../Core/Series/Series.js';
+const {
+    seriesTypes: {
+        column: ColumnSeries
+    }
+} = BaseSeries;
+import U from '../../Core/Utilities.js';
+const { isNumber } = U;
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+class FlagsPoint extends ColumnSeries.prototype.pointClass {
+
+    /* *
+     *
+     *  Properties
+     *
+     * */
+
+    public _y?: number;
+
+    public anchorX?: number;
+
+    public options: FlagsPointOptions = void 0 as any;
+
+    public series: FlagsSeries = void 0 as any;
+
+    public fillColor?: ColorType;
+
+    public lineWidth?: number;
+
+    public raised?: boolean;
+
+    public stackIndex?: number;
+
+    public style?: CSSObject;
+
+    /* *
+     *
+     *  Functions
+     *
+     * */
+
+    /* eslint-disable valid-jsdoc */
+
+    /**
+     * @private
+     */
+    public isValid(): boolean {
+        // #9233 - Prevent from treating flags as null points (even if
+        // they have no y values defined).
+        return isNumber(this.y) || typeof this.y === 'undefined';
+    }
+
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default FlagsPoint;
